Reviewed chapter2

+++ chapter2/f.c
@@ -13,11 +13,11 @@
 
 void main()
 {
-	int t, v;
+	float t, v;
 	float wcf;
 	
 	printf("Enter the temperature and velocity\n");
-	scanf("%d%d", &t, &v);
+	scanf("%f%f", &t, &v);
 	
 	wcf = 35.74 + (0.6215 * t) + ((0.4275 * t) - 35.75) * pow(v, 0.16);
